wire: add deserialization+serialization of witness transactions, update tests
This commit introduces a new transaction related struct: TxWitness. A
design choice has been made to includes TxWitness within a TxIn, rather
than as a new top level fields within a MsgTx (such as TxOut, TxIn).
The rationale behind this choice is that, logically, a particular
TxWitness _belongs_ to an input. Therefore, from the point of view of
developers, it is more natural to package a witness with its inputs
within the high level data structure. The encoding which places all
witness after TxOut’s, but before the Version is still maintained in
order to be compatible with BIP0141.