wire: add new WTxSha function to calculate a transaction's wtxid
A new transaction identifier: wtxid has been defined as part of
BIP0141. The wtxid is used within the new witness commitment within the
coinbase in place of a regular wtxid.

The wtxid of a coinbase transaction is al zeroes. The
blockchain.IsCoinbase function has been replicated here in order to
avoid an import cycle.