wire: introduce wire.MessageEncoding along with new read/write funcs
This commit introduces a degree of freedom to the wire.Message
interface. BtcEncode, and BtcDecode now additionally take a
wire.MessageEncoding parameter. This allows callers to specify which
encoding format should be expected, or used when reading/writing a
message.

The motivating use-case for this change is the new transaction encoding
format described within BIP0144. Due to this new encoding format, a
peer may wish to selectively use one encoding if the peer advertises
support for the new BIP, and the old legacy encoding if not.

Two new methods have been added wire.WriteMessageWithEncodingN, and
wire.ReadMessageWithEncodingN. These two methods allow the caller to
control which encoding should be expected/used by the respective
functions.

All tests have been updated accordingly to account for the new
interface.