txscript: add segwit sighash calculation and HashCache integration
This commit adds the new segwit sighash digest calculation algorithm
introduced as part of BIP0143. This functionality has been exposed as
an internal function to be used by the new signing logic (which
includes the input amount).

Additionally, a new caching structure, the HashCache has been
introduced. The HashCache is a thread safe cache which allows multiple
goroutines to safety modify, and query the partial sighashes will allow
the algorithmic improvement in the digest calculation.