txscript+blockchain: optimize HashCache usage
Previously the transaction’s txid would be re-computed twice: once when
checking for existence within the hash cache, and second when adding
the transactions sighash misstate to the hashcahe. With very large
transactions with thousands of inputs, this resulted in a large
performance hit during validation.

Instead, we now only compute the transaction’s txid once. The
CointainsHashes, and GetSigHashes functions of the HashCache now take
the txid directly instead of the transaction which allows for re-use of
a prior computed txid value.

Within txscript.Engine, rather than passing in the entire HashCache,
only the sighash midstate for that particular transaction is fed in.
This allows us to avoid the double hashing again when checking for
existence.