multi: when encoding/decoding make segwit the default not exception
This commit removes the previous msgTx/msgBlock.SerializeWitness method
in favor of simply making Serialize encode witness data if it’s
available. Additionally, msgTx/msgBlock.SerializeSize now includes the
size of the witness data by default. A new method SerializeSizeStripped
has been introduced to calculate the serialize size excluding the
witness data.

This change was made in order to ease the developer experience when
dealing with blocks/transactions with witness data. As a bonus, this
caught several dormant bugs.