default to using Median Time Past during block validation (place holder commit)
This commit defaults to using the BIP 113 median time past rules when
validating blocks from peers, or those sent over the RPC interface.

This commit is a place holder, as this behavior should be guarded by a
check to the versionbits state machine in order to determine if the
rule has been activated and locked-in.