rpctest: add integration test for BIP 113
This commit adds an integration test using the new rpctest package to
ensure the implementation of BIP 113 functions as expected.

The test excursus that transactions which are “final” according to the
latest block-time but not the current MTP are rejected fro the mempool.
Additionally we check that once a transaction is final according to the
current MTP, it’s accepted into the mempool and also mined into the
following block.